# Adds a simple benchmarking tool.