… test/bench/many.fla
… test/bench/many.swf