When making changes to performance critical code, we should run this benchmarking tool. The first time you run:

```
node utils/bench.js
```

A file named `baseline.json` is saved in the Shumway root directory. Subsequent runs of the benchmarking tool will compare against this baseline. If you want to clear the baseline, you can simply delete the `baseline.json` file.

The tool prints out GitHub compatible Markdown tables, so paste them in your PRs:

|     Metric    | Baseline |   New   | Difference |   %   |                                    Command                                   | 
|:--------------|:---------|:--------|:-----------|------:|:-----------------------------------------------------------------------------|
| HelloWorld    | 201ms    | 212ms   | 10ms       | 5.05% | js build/ts/shell.js -r -x test/avm2/shumway/hello.abc                       | 
| Many w/o Pump | 1,741ms  | 1,779ms | 38ms       | 2.20% | js build/ts/shell.js -x test/bench/many.swf -r -det -fc 30 -enablePump false | 
| Many          | 2,144ms  | 2,234ms | 90ms       | 4.19% | js build/ts/shell.js -x test/bench/many.swf -r -det -fc 30                   | 

Like so.

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/mozilla/shumway/2368)
<!-- Reviewable:end -->

To get more consistent results, play with the iteration count. It's currently set to 10.