@mbebenita 2015-09-03
We need to make sure benchmarks are not run when CPU utilization is high. We now prevent them from running until it dips below 5%.

|    Key    |                   Value                   | 
|:----------|:------------------------------------------|
| CPU Model | Intel(R) Core(TM) i7-4980HQ CPU @ 2.80GHz | 
| CPU Speed | 2800                                      | 
| CPU Count | 8                                         | 
| Idle Time | 95.31%                                    | 


|     Metric    | Baseline |   New   | Difference |    %   |                                    Command                                   | 
|:--------------|:---------|:--------|:-----------|-------:|:-----------------------------------------------------------------------------|
| HelloWorld    | 200ms    | 188ms   | -12ms      | -6.17% | js build/ts/shell.js -r -x test/avm2/shumway/hello.abc                       | 
| Many w/o Pump | 1,776ms  | 1,754ms | -21ms      | -1.20% | js build/ts/shell.js -x test/bench/many.swf -r -det -fc 30 -enablePump false | 
| Many          | 1,766ms  | 1,744ms | -22ms      | -1.25% | js build/ts/shell.js -x test/bench/many.swf -r -det -fc 30                   |