@mbebenita 2015-09-03
Adds standard deviation.

|     Metric    | Baseline |   New   |  Std | Difference |    %   |                                    Command                                   | 
|:--------------|:---------|:--------|:-----|:-----------|-------:|:-----------------------------------------------------------------------------|
| HelloWorld    | 194ms    | 187ms   | 3ms  | -7ms       | -3.44% | js build/ts/shell.js -r -x test/avm2/shumway/hello.abc                       | 
| Many w/o Pump | 1,815ms  | 1,733ms | 45ms | -82ms      | -4.50% | js build/ts/shell.js -x test/bench/many.swf -r -det -fc 30 -enablePump false | 
| Many          | 1,794ms  | 1,731ms | 19ms | -63ms      | -3.52% | js build/ts/shell.js -x test/bench/many.swf -r -det -fc 30                   |